code: attempt to add a second step to hinding a modal
Currently, hiding a modal immediately resets the global "active modal"
value. This works when we don't have a transition on the modal, but when
we do, then we have to ensure that the data in the modal is visible
while it transitions out, otherwise we have a weird change during the
transition.