code: fix bug with adding reports to a dashboard
When a newly created report was added to a dashboard, it was being added
without an id, which meant when it was being saved to the server, the
server was ignoring it since it couldn't match that with a report that
was on the server. Instead, append the response from the create request
which shouldn't result in this issue.