fix(analytics-control-bar-filter): fix bug in analytics control bar filter
I think rendering the backdrop in a portal for the analytics control bar
filter works well in the storybook, but because the control bar is not
rendered at the top level in the actual page, using a portal causes the
stacking contexts to be different. So, instead, I stopped rendering the
backdrop in a portal and it fixed it.