bug(rx-stores): ensure that actions aren't dispatched twice
If an action was dispatched with rxDispatch, it seemed to dispatch twice because it would be
dispatched t oredux, and then the redux middleware would then dispatch it again as it had no way to
tell if the action was dispatched from rxDispatch or from the redux store's store.dispatch.

To fix this, I added a key to each action that is dispatched into redux called legacyReduxAction and
then am checkng this key to see where the action came from when re-dispatching it in the middleware.